# Description

I have tried to implement a dendrogram plot for the plot_missing module. Bokeh does not have an inbuilt dendrogram functionality. **The approach I have used is as follows:**
1) Use scipy and matplotlib to render the dendrogram image and Generate a RGBA numpy array of the image. This is done in the compute.py of the missing module.
3) Finally render the RGBA array using Bokeh, in render.py of the missing module.

Currently the resolution of the plot is **not** optimal and the matplotlib image also gets rendered below the bokeh image. I have tried to remove the matplotlib image but keeps showing up. I have attached a screenshot below. 

![Screen Shot 2020-07-24 at 10 55 31 PM](https://user-images.githubusercontent.com/59293179/88450194-00fc1b80-ce02-11ea-8d8c-ca1349c97192.png)


# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration.

I have tested this on the titanic dataset and a few other datasets, it seems to work but the resolution is again not optimal and matplotlib image shows below.

# Snapshots:

Include snapshots for easier review.

# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have already squashed the commits and make the commit message conform to the project standard.
- [x] I have already marked the commit with "BREAKING CHANGE" or "Fixes #" if needed.
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules