@peterirani 2020-07-28
Update - trying to render this using Bokeh and Scipy.