@peterirani 2020-07-31
The dendrogram using Bokeh renders fine but if we re-run the cell, matplotlib image also gets rendered. 

** first time when you run the cell**
<img width="855" alt="Screen Shot 2020-07-30 at 10 56 16 PM" src="https://user-images.githubusercontent.com/59293179/89005763-1c5f9e80-d2ba-11ea-9c64-3ec7b67447b9.png">

**once you re-run the cell
<img width="690" alt="Screen Shot 2020-07-30 at 10 56 28 PM" src="https://user-images.githubusercontent.com/59293179/89005801-300b0500-d2ba-11ea-8f9c-494aff6fbe8a.png">

I can't figure out why the image (I think it is matplotlib) is being rendered.