@peterirani 2020-08-01
Thanks, Brandon `no_plot` works! I will work on the hover tips, and 50 columns issues now.