@peterirani 2020-08-06
Benchmark timing before and after adding missing_dendrogram plot. 

**BEFORE**
<img width="902" alt="Screen Shot 2020-08-06 at 11 09 17 AM" src="https://user-images.githubusercontent.com/59293179/89570472-19ddd700-d7db-11ea-8715-5cdcecab0abc.png">


**AFTER**
<img width="867" alt="Screen Shot 2020-08-06 at 11 04 51 AM" src="https://user-images.githubusercontent.com/59293179/89570517-29f5b680-d7db-11ea-8fa8-b18963213b77.png">