@dovahcrow 2020-08-07
Add some optimizations. On my machine, 5M rows adult dataset
Before:
CPU times: user 10.1 s, sys: 2.06 s, total: 12.2 s
Wall time: 7.46 s
After:
CPU times: user 4.94 s, sys: 768 ms, total: 5.71 s
Wall time: 3.07 s

@peterirani I think your commit message is wrong. There should not be a space between ")" and ":". Otherwise, the release note will not be shown correctly as mentioned on discord 2 weeks ago. I've revised the commit message.