@peterirani 2020-08-07
Thank you Young, Brandon and I are working on the documentation for dendrogram.